snap: Add FRR components to MicroOVN.
This change adds `frr` part to the MicroOVN snap with following
FRR components enabled:
 * Zebra daemon
 * BGP daemon
 * VTYSH CLI application for communication with daemons

Due to issues with privlege separation module, we are unable to use
FRR from ubuntu repositories at this time. Biggest issue is that
FRR specifies allowed users and groups at build time [0]. During
the runtime, `zebra` daemon tries to make sure that the user is part
of the `frrvty` group but this fails due to the sctrict confiment.
As a result, we have to build the FRR from sources ourselves.

Note: BGP daemon sidesteps this issue by using -S option [1].

[0] https://docs.frrouting.org/en/stable-10.0/installation.html#cmdoption-configure-enable-vty-group
[1] https://docs.frrouting.org/en/stable-10.0/bgp.html#cmdoption-bgpd-S

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>