bgp: Resource cleanup on bgp disable
When users disable bgp service, we need to clean up
resources that were created for BGP redirection. This
change utilises "external-ids" column on various
OVN/OVS resources to tag them when they are created.
During the teardown, we can then find those resources
based on the external-ids and remove them.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>