bgp: Re-apply BGP config on daemon restart.
While most of the resources are managed by OVN and on service (or
machine) restart will be re-instated, some of the resources need
manual attention.

Specifically, the interfaces that are used for BGP redirection,
need to be re-assigned to the VRF and they need to have their
IP address reconfigured.

In the future implementations, it'd be best if we could offload
management of interfaces to some other tool, for example, Netplan.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>