bgp: Enable IPv4 over IPv6 routing.
By applying patches proposed to upstream OVN that enable
IPv4 routes with IPv6 nexthop [0], we can experimentally
enable this functionality in the MicroOVN as well.

When user enables BGP in MicroOVN, IPv4 routes leaked
by the OVN will be advertised by the FRR. The next hop
address for these routes will be IPv6 Link Local address
of the OVN router that originally announced the route.

[0] https://patchwork.ozlabs.org/project/ovn/list/?series=410497&archive=both&state=*

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>