Changes microovn bgp to use a veth pair
Adds netplan integration, which is so far only used to generate a veth
pair, I would like to expand it to include the interface bridge created
and also rework how you build a netplan yaml file.

The main change is this avoids remastering the eth-bgp interface to the
vrf, we instead have a veth peer with one side on ovn and the other on
the vrf.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>